requirements.txt
python-docx
mammoth
pyyaml
fastapi
uvicorn
grpcio
grpcio-tools
